pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "google-genai>=1.73.1",
    "ipykernel>=7.2.0",
    "jinja2>=3.1.6",
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "notebook>=7.5.5",
    "numpy>=2.4.4",
    "opencv-python-headless>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "python-multipart>=0.0.26",
    "scikit-learn>=1.8.0",
    "tensorflow-cpu>=2.21.0",
    "uvicorn[standard]>=0.46.0",
]
